// console/src/generated/sdk/SettingsDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { InstanceSettingsDto } from "./InstanceSettingsDto";
import type { LoggingSettingsDto } from "./LoggingSettingsDto";

export type SettingsDto = { instance: InstanceSettingsDto, logging: LoggingSettingsDto, };